studio/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
httpx
aiofiles
